kleinanzeigen-bot/services/telegram-bot/requirements.txt
python-telegram-bot==21.*
redis==5.*
psycopg2-binary==2.9.*
httpx==0.28.*
Pillow==11.*
